refactor: processing node redirect

+++ server/internal/imagorprovider/processor_cache.go
@@ -4,10 +4,10 @@ import "github.com/cshum/imagor-studio/server/internal/config"
 
 const defaultProcessorCacheSizeBytes = 200 * 1024 * 1024
 
-func processorCacheSizeBytes(cfg *config.Config) int {
+func processorCacheSizeBytes(cfg *config.Config) int64 {
 	cacheSizeBytes := int64(defaultProcessorCacheSizeBytes)
 	if cfg != nil && cfg.ImagorCacheSizeBytes > 0 {
 		cacheSizeBytes = cfg.ImagorCacheSizeBytes
 	}
-	return int(cacheSizeBytes)
+	return cacheSizeBytes
 }
